+++ src/clj/datasplash/api.clj
@@ -40,6 +40,7 @@
 (intern *ns* (with-meta 'group-by (meta #'dt/dgroup-by)) @#'dt/dgroup-by)
 (intern *ns* (with-meta 'make-pipeline (meta #'dt/make-pipeline)) @#'dt/make-pipeline)
 (intern *ns* (with-meta 'run-pipeline (meta #'dt/run-pipeline)) @#'dt/run-pipeline)
+(intern *ns* (with-meta 'wait-pipeline-result (meta #'dt/wait-pipeline-result)) @#'dt/wait-pipeline-result)
 (intern *ns* (with-meta 'cogroup-by (meta #'dt/cogroup-by)) @#'dt/cogroup-by)
 (intern *ns* (with-meta 'join-by (meta #'dt/join-by)) @#'dt/join-by)
 (intern *ns* (with-meta 'distinct (meta #'dt/ddistinct)) @#'dt/ddistinct)
@@ -107,7 +108,6 @@
 (intern *ns* (with-meta 'write-edn-file (meta #'dt/write-edn-file)) @#'dt/write-edn-file)
 (intern *ns* (with-meta 'write-json-file (meta #'dt/write-json-file)) @#'dt/write-json-file)
 (intern *ns* (with-meta 'generate-input (meta #'dt/generate-input)) @#'dt/generate-input)
-(intern *ns* (with-meta 'walk-gcs-tree (meta #'dt/walk-gcs-tree)) @#'dt/walk-gcs-tree)
 
 ;;;;;;;;;;;;;
 ;; Formats ;;