+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 #Changes for Datasplash
 
+## v0.5.0 (2017-09-18)
+
+### Changes
+
+- Update to v2.x Dataflow version (Apache Beam)
+- Add new streaming features (custom timestamp in PubSub #30 (thanks to @RoIT), BigQuery streaming writes)
+- Add compressed TextIO output #31 (thanks to @RoIT)
+- Add FileNamePolicy compatibility for TextIO
+
 ## v0.4.1 (2017-03-09)
 
 **[compare](https://github.com/ngrunwald/datasplash/compare/v0.4.0...v0.4.1)**